fix management feature

+++ src/com/qushop/order/service/Order_detailService.java
@@ -16,7 +16,7 @@ public interface Order_detailService {
 	 */
 	public List<Order_detail> getOrderdetailByMethod(int type,Oper oper,String...params);
 	
-	public Order_detail getOrderdetail(String productId, String orderId,int orderType) ;
+	public Order_detail getOrderdetail(String productId, String orderId,short orderType) ;
 	
 	public boolean deleteOrderDetail(String orderIds);
